rename: move some post content component to common
