+++ src/components/common/checkbox.tsx
@@ -1,8 +1,8 @@
-import { postContentCheckbox } from '@/styles';
+import { checkboxClassName } from '@/styles';
 import { InputHTMLAttributes } from 'react';
 
 export default function Checkbox({
 	...props
 }: InputHTMLAttributes<HTMLInputElement>) {
-	return <input type='checkbox' {...props} className={postContentCheckbox} />;
+	return <input type='checkbox' {...props} className={checkboxClassName} />;
 }
