+++ src/components/common/index.ts
@@ -1,5 +1,9 @@
 import BackLink from './backLink';
+import Checkbox from './checkbox';
 import Heading from './heading';
+import Image from './image';
+import Link from './link';
+import List from './list';
 import Text from './text';
 
-export { BackLink, Heading, Text };
+export { BackLink, Checkbox, Heading, Image, Link, List, Text };
