+++ src/components/post/elements/index.ts
@@ -1,23 +1,5 @@
-import Checkbox from './checkbox';
-import Heading from './heading';
 import HrLine from './hrLine';
-import Image from './image';
-import Link from './link';
-import List from './list';
 import Quote from './quote';
 import { Table, Td, Th } from './table';
-import Text from './text';
 
-export {
-	Checkbox,
-	Heading,
-	HrLine,
-	Image,
-	Link,
-	List,
-	Quote,
-	Table,
-	Td,
-	Text,
-	Th,
-};
+export { HrLine, Quote, Table, Td, Th };
