--- src/components/post/elements/text.tsx
@@ -1,5 +1,0 @@
-import { PropsWithChildren } from 'react';
-
-export default function Text({ children }: PropsWithChildren) {
-	return <p>{children}</p>;
-}
